LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-shadow-expected.txt


FAIL @property rules in shadow trees should have no effect assert_equals: expected "2px" but got " calc(1px + 1px)"
